+++ CHANGELOG.md
@@ -9,6 +9,7 @@ This file is used to list changes made in each version of the gem.
 - Split DockerGenerator in multiple sub-generators to version each one individually.
 - Merge all keepfiles in one.
 - Wrap health_check initializer in a reloader.to_prepare block.
+- Always use latest ruby in dockeruby.
 
 # 1.5.1